+++ R/scrapers.R
@@ -2,6 +2,8 @@
 #'
 #' Substitute all German "Umlaute"
 #' @param x A character vector.
+#' @return A character vector with German umlauts replaced by their ASCII
+#' equivalents (e.g. "oe" for "ö").
 #' @keywords internal
 sanitize_strings <- function(x) {
 
@@ -51,6 +53,7 @@ sanitize_sonstige <- function(x) {
 #' Removes all characters that are not in [0-9].
 #' @param x A character vector.
 #' @param decimal Logical flag, indicating if x has a decimal separator
+#' @return A numeric vector with non-numeric characters removed.
 #' @keywords internal
 extract_num <- function(x, decimal=TRUE) {
 
@@ -67,6 +70,8 @@ extract_num <- function(x, decimal=TRUE) {
 #'
 #' @param df A data frame with party names with special characters that need
 #' to be sanitized.
+#' @return The input data frame with column names converted to lowercase ASCII
+#' and party name abbreviations standardized.
 #' @keywords internal
 sanitize_colnames <- function(df) {
 
@@ -84,6 +89,7 @@ sanitize_colnames <- function(df) {
 #' Try call of read_html that throws an error if the url cannot be resolved
 #'
 #' @param url http-address that should be scraped.
+#' @return An \code{xml_document} object as returned by \code{xml2::read_html}.
 #' @importFrom xml2 read_html
 try_readHTML <- function(url) {
 
@@ -104,24 +110,22 @@ try_readHTML <- function(url) {
 #' @rdname scrape
 #' @param address http-address from which tables should be scraped.
 #' @param parties A character vector containing names of parties to collapse.
+#' @return A tibble with one row per survey date and columns for date,
+#' respondents, and one column per party containing the percentage of votes.
 #' @import rvest dplyr
 #' @importFrom lubridate dmy
 #' @importFrom stringr str_sub
 #' @importFrom rlang .data
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' library(coalitions)
 #' library(dplyr)
-#' # select a polling agency from .pollster_df that should be scraped ...
-#' coalitions:::.pollster_df
-#' # ... here we choose Forsa
-#' address <- coalitions:::.pollster_df %>% filter(pollster == "forsa") %>% pull(address)
-#' scrape_wahlrecht(address = address) %>% slice(1:5)
+#' scrape_wahlrecht() %>% slice(1:5)
 #' }
 #' @export
 scrape_wahlrecht <- function(
   address = "https://www.wahlrecht.de/umfragen/emnid.htm",
-  parties = c("CDU", "SPD", "GRUENE", "FDP", "LINKE", "PIRATEN", "FW", "AFD",
+  parties = c("CDU", "SPD", "GRUENE", "FDP", "LINKE", "PIRATEN", "AFD", "BSW",
     "SONSTIGE")) {
 
   atab <- try_readHTML(address) %>%
@@ -141,27 +145,27 @@ scrape_wahlrecht <- function(
   atab           <- atab[ind_row_remove, ]
   atab           <- atab[-nrow(atab), ]
   colnames(atab) <- c("Datum", colnames(atab)[-1])
-  
+
   if (any(nchar(atab$Sonstige) > 6)) {
     # correct the 'Sonstige' column if it contains information on
     # one party + other parties (see issue #138)
     weird_rows <- which(nchar(atab$Sonstige) > 6)
     if (length(weird_rows) > 0) {
       for (row in weird_rows) {
-        entry  <- atab$Sonstige[row] %>% 
+        entry  <- atab$Sonstige[row] %>%
           gsub(pattern = ",", replacement = ".")
         # include '0' instead of decimal commas s.t. gregexpr correctly finds the beginning of each decimal number
         entry_gregexpr <- gsub(atab$Sonstige[row], pattern = ",", replacement = ".")
-        shares <- entry_gregexpr %>% 
-          gregexpr("[[:digit:]]+", .) %>% 
-          regmatches(entry, .) %>% 
-          unlist() %>% 
+        shares <- entry_gregexpr %>%
+          gregexpr("[[:digit:]]+", .) %>%
+          regmatches(entry, .) %>%
+          unlist() %>%
           as.numeric()
         atab$Sonstige[row] <- paste(sum(shares), "%")
       }
     }
   }
-  
+
   ind.empty      <- sapply(atab, function(z) { all(z == "") | all(is.na(z)) } ) |
     sapply(colnames(atab), function(z) z == "") |
     sapply(colnames(atab), function(z) is.na(z))
@@ -171,13 +175,13 @@ scrape_wahlrecht <- function(
   parties <- colnames(atab)[colnames(atab) %in% tolower(parties)]
   # transform percentage string to numerics
   atab <- atab %>%
-    mutate_at(c(parties), extract_num) %>%
-    mutate_at("befragte", extract_num, decimal = FALSE)
+    mutate(across(all_of(parties), extract_num)) %>%
+    mutate(across(all_of("befragte"), ~extract_num(.x, decimal = FALSE)))
 
   atab <- mutate(atab, datum = dmy(.data$datum))
   atab <- mutate(atab, zeitraum = case_when(nchar(zeitraum) == 6 ~ paste0(zeitraum, "-", zeitraum),
                                             TRUE                 ~ zeitraum))
-  
+
   atab <- atab %>%
     filter(.data$zeitraum != "Bundestagswahl",
            !grepl("\\?", .data$zeitraum)) %>%
@@ -189,16 +193,28 @@ scrape_wahlrecht <- function(
       end   = case_when(
         is.na(end) ~ start,
         TRUE ~ end))
-
+  
+  # if present add FW to others
+  if ("fw" %in% colnames(atab)) {
+      atab <- atab %>%
+      mutate(across(all_of("fw"), extract_num))
+    if (any(!is.na(atab$fw))) {
+      atab <- atab %>%
+        mutate(
+          fw = tidyr::replace_na(.data$fw, 0),
+          sonstige = .data$sonstige + .data$fw)
+    }
+  }
+  
   atab <- atab %>%
     mutate(total = rowSums(atab[, parties], na.rm = TRUE)) %>%
     filter(.data$total == 100, !is.na(.data$befragte), !is.na(.data$datum)) %>%
-    select(one_of(c("datum", "start", "end", parties, "befragte")))
-  
+    select(any_of(c("datum", "start", "end", parties, "befragte")))
+
   # remove potential duplicate entries
   atab <- atab %>%
     group_by(.data$datum) %>% slice(1) %>% ungroup()
-  
+
   colnames(atab) <- prettify_strings(
     colnames(atab),
     current = .trans_df$german,
@@ -212,7 +228,7 @@ scrape_wahlrecht <- function(
 #'
 #' @rdname get_surveys
 #' @param country Choose country from which surveys should be scraped.
-#' Currently \code{"DE"} (Germany) and \code{"AT"} (Austria) are supported.
+#' Currently \code{"DE"} (Germany) is supported.
 #' @import dplyr
 #' @importFrom purrr map
 #' @return Nested tibble. When fully unnested, the dataset contains the following
@@ -229,26 +245,21 @@ scrape_wahlrecht <- function(
 #'   \item{votes}{Number of respondents that chose the party.}
 #' }
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' library(coalitions)
-#' # scrape data for the German federal election
-#' # get_surveys()
+#' get_surveys()
 #' }
 #' @export
-get_surveys <- function(country = c("DE", "AT")) {
+get_surveys <- function(country = "DE") {
 
   country <- match.arg(country)
 
-  if (country == "DE") {
-    surveys <- .pollster_df %>%
+  surveys <- .pollster_df %>%
     mutate(
       surveys = map(.data$address, scrape_wahlrecht),
-      surveys = map(.x = surveys, collapse_parties)) %>%
-    select(-one_of("address"))
-  }
-  if (country == "AT") {
-    surveys <- scrape_austria()
-  }
+      surveys = map(.x = surveys, collapse_parties,
+                    parties = c("cdu", "spd", "greens", "fdp", "left", "pirates", "afd", "bsw", "others"))) %>%
+    select(-any_of("address"))
 
   surveys
 
@@ -286,8 +297,8 @@ scrape_by <- function(
   parties <- colnames(atab)[colnames(atab) %in% tolower(parties)]
   # transform percentage string to numerics
   atab <- atab %>%
-    mutate_at(c(parties), extract_num) %>%
-    mutate_at("befragte", extract_num, decimal = FALSE)
+    mutate(across(all_of(parties), extract_num)) %>%
+    mutate(across(all_of("befragte"), ~extract_num(.x, decimal = FALSE)))
 
   atab <- atab %>%
     mutate(
@@ -298,7 +309,7 @@ scrape_by <- function(
   atab <- atab %>%
     mutate(total = rowSums(atab[, parties], na.rm = TRUE)) %>%
     filter(.data$total == 100, !is.na(.data$befragte), !is.na(.data$datum)) %>%
-    select(one_of(c("institut", "datum", "start", "end", parties, "befragte"))) %>%
+    select(any_of(c("institut", "datum", "start", "end", parties, "befragte"))) %>%
     rename(pollster = "institut") %>%
     mutate(
       pollster = tolower(.data$pollster),
@@ -324,7 +335,7 @@ get_surveys_by <- function() {
   by <- scrape_by()
   by %>%
     collapse_parties(parties = c("csu","spd","greens","fdp","left","pirates","fw","afd","others")) %>%
-    nest(surveys = -one_of("pollster"))
+    nest(surveys = -any_of("pollster"))
 
 }
 
@@ -358,8 +369,8 @@ scrape_rp <- function(
   # transform percentage string to numerics
   atab <- atab %>%
     mutate(sonstige = sanitize_sonstige(.data$sonstige)) %>%
-    mutate_at(c(parties), extract_num) %>%
-    mutate_at("befragte", extract_num, decimal = FALSE)
+    mutate(across(all_of(parties), extract_num)) %>%
+    mutate(across(all_of("befragte"), ~extract_num(.x, decimal = FALSE)))
 
   atab <- atab %>%
     mutate(
@@ -370,7 +381,7 @@ scrape_rp <- function(
   atab <- atab %>%
     mutate(total = rowSums(atab[, parties], na.rm = TRUE)) %>%
     filter(.data$total == 100, !is.na(.data$befragte), !is.na(.data$datum)) %>%
-    select(one_of(c("institut", "datum", "start", "end", parties, "befragte"))) %>%
+    select(any_of(c("institut", "datum", "start", "end", parties, "befragte"))) %>%
     rename(pollster = "institut") %>%
     mutate(
       pollster = tolower(.data$pollster),
@@ -395,7 +406,7 @@ get_surveys_rp <- function() {
   rp <- scrape_rp()
   rp %>%
     collapse_parties() %>%
-    nest(surveys = -one_of("pollster"))
+    nest(surveys = -any_of("pollster"))
 
 }
 
@@ -406,7 +417,7 @@ get_surveys_rp <- function() {
 #' @param ind_row_remove Negative vector of rows that will be skipped at the beginning.
 #' @export
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' # Niedersachsen
 #' scrape_ltw() %>% slice(1:5)
 #' # Hessen
@@ -415,7 +426,7 @@ get_surveys_rp <- function() {
 #' }
 scrape_ltw <- function(
   address = "https://www.wahlrecht.de/umfragen/landtage/niedersachsen.htm",
-  parties = c("CDU", "SPD", "GRUENE", "FDP", "LINKE", "PIRATEN", "FW", "AFD",
+  parties = c("CDU", "SPD", "GRUENE", "FDP", "LINKE", "PIRATEN", "FW", "AFD", "BSW",
     "SONSTIGE"),
   ind_row_remove = -c(1:2)) {
 
@@ -437,8 +448,8 @@ scrape_ltw <- function(
   # transform percentage string to numerics
   atab <- atab %>%
     mutate(sonstige = sanitize_sonstige(.data$sonstige)) %>%
-    mutate_at(c(parties), extract_num) %>%
-    mutate_at("befragte", extract_num, decimal = FALSE)
+    mutate(across(all_of(parties), extract_num)) %>%
+    mutate(across(all_of("befragte"), ~extract_num(.x, decimal = FALSE)))
 
   atab <- atab %>%
     mutate(
@@ -449,7 +460,7 @@ scrape_ltw <- function(
   atab <- atab %>%
     mutate(total = rowSums(atab[, parties], na.rm = TRUE)) %>%
     filter(.data$total == 100, !is.na(.data$befragte), !is.na(.data$datum)) %>%
-    select(one_of(c("institut", "datum", "start", "end", parties, "befragte"))) %>%
+    select(any_of(c("institut", "datum", "start", "end", parties, "befragte"))) %>%
     rename(pollster = "institut") %>%
       mutate(
         pollster = tolower(.data$pollster),
@@ -475,7 +486,7 @@ get_surveys_nds <- function() {
   nds <- scrape_ltw()
   nds %>%
     collapse_parties() %>%
-    nest(surveys = -one_of("pollster"))
+    nest(surveys = -any_of("pollster"))
 
 }
 
@@ -488,7 +499,7 @@ get_surveys_saxony <- function() {
     "https://www.wahlrecht.de/umfragen/landtage/sachsen.htm",
     ind_row_remove = -1)
   saxony %>% collapse_parties() %>%
-    nest(surveys = -one_of("pollster"))
+    nest(surveys = -any_of("pollster"))
 
 }
 
@@ -498,7 +509,7 @@ get_surveys_brb <- function() {
 
   brb <- scrape_ltw("https://www.wahlrecht.de/umfragen/landtage/brandenburg.htm")
   brb %>% collapse_parties() %>%
-    nest(surveys = -one_of("pollster"))
+    nest(surveys = -any_of("pollster"))
 
 }
 
@@ -510,68 +521,6 @@ get_surveys_thuringen <- function() {
     "https://www.wahlrecht.de/umfragen/landtage/thueringen.htm",
     ind_row_remove = -1)
   thuringen %>% collapse_parties() %>%
-    nest(surveys = -one_of("pollster"))
+    nest(surveys = -any_of("pollster"))
 
 }
-
-
-#' Import Austrian survey results
-#'
-#' Reads JSON file from neuwal.com and performs some preprocessing to bring
-#' data into standardized format. Returns a nested tibble.
-#'
-#'
-#' @param address URL of the JSON file.
-#' @import dplyr
-#' @importFrom tidyr nest unnest
-#' @importFrom purrr map map_dfr flatten_dfc
-#' @importFrom jsonlite fromJSON
-#' @importFrom RCurl getURL
-#' @importFrom lubridate dmy
-#' @importFrom rlang UQS syms .data
-#' @importFrom stringr str_replace
-#' @export
-scrape_austria <- function(
-  address = "https://neuwal.com/wahlumfragen/data/neuwal-wahlumfragen-user.json") {
-
-  raw_input <- getURL(address)
-  raw_input <- str_replace(raw_input, '\\"\\"(.*)\\"\\",', "\"'\\1'\",") # fix for double double-quote bug
-  aut_list <- fromJSON(raw_input)
-  out_df <- as_tibble(aut_list[[1]]) %>%
-    filter(.data$regionID == 1)
-  party <- out_df %>%
-    select(one_of("id"), contains("Party")) %>%
-    gather("key", "party", contains("Party")) %>%
-    arrange(desc(.data$id)) %>%
-    select(-.data$key)
-  party <- party %>%
-    nest(data = .data$party) %>%
-    mutate(data = map(.data$data, ~rbind(.x, data.frame(party = "Others")))) %>%
-    unnest("data")
-  percent <- out_df %>%
-    select(one_of(c("id", "n")), contains("Value")) %>%
-    gather("key", "percent", contains("Value")) %>%
-    arrange(desc(.data$id)) %>%
-    mutate(percent = as.numeric(.data$percent)) %>%
-    select(-one_of("key"))
-  percent <- percent %>% nest(data = .data$percent) %>%
-    mutate(data = map(.data$data, ~rbind(.x, data.frame(percent = 100 - sum(.x$percent))))) %>%
-    unnest("data") %>%
-    mutate(votes   = .data$n * .data$percent / 100) %>%
-    select(-n)
-  pp <- cbind(party, percent[, -1]) %>%
-    nest(survey = -.data$id)
-
-  out_df <- out_df %>%
-    select(one_of(c("id", "institut", "n", "datum"))) %>%
-    rename(pollster = "institut", date = "datum", respondents = "n")
-  out_df <- out_df %>% left_join(pp)
-  out_df %>%
-    mutate(
-      date  = as.Date(.data$date),
-      start = .data$date,
-      end   = .data$date) %>%
-    select(one_of(c("pollster", "date", "start", "end", "respondents", "survey"))) %>%
-    nest(surveys = -.data$pollster)
-
-  }