+++ R/seat-distribution.R
@@ -21,7 +21,7 @@
 #' library(coalitions)
 #' library(dplyr)
 #' # get the latest survey for a sample of German federal election polls
-#' surveys <- get_latest(surveys_sample)
+#' surveys <- get_latest(surveys_sample) %>% ungroup() %>% slice(1)
 #' # simulate 100 seat distributions
 #' surveys <- surveys %>% mutate(draws = purrr::map(survey, draw_from_posterior, nsim = 100),
 #'                               seats = purrr::map2(draws, survey, get_seats))
@@ -47,14 +47,14 @@ get_seats <- function(
 
   dirichlet.draws %>%
     mutate(sim = row_number()) %>%
-    gather("party", "percent", -.data$sim) %>%
+    pivot_longer(names_to = "party", values_to = "percent", cols = !all_of("sim")) %>%
     arrange(.data$sim) %>%
     mutate(votes = .data$percent * samplesize) %>%
     filter(.data$party != others & (.data$percent >= hurdle | .data$party == "ssw")) %>%
     group_by(.data$sim) %>%
     mutate(seats = distrib.fun(.data$votes, .data$party, ...)) %>%
     ungroup() %>%
-    select(one_of(c("sim", "party", "seats")))
+    select(any_of(c("sim", "party", "seats")))
 
 }
 
@@ -64,12 +64,14 @@ get_seats <- function(
 #' @inheritParams get_seats
 #' @param epsilon Percentages should add up to 1. If they do not, within accuracy
 #' of \code{epsilon}, an error is thrown.
+#' @return A data frame with the same structure as \code{survey} but with
+#' parties below the hurdle removed and vote percentages renormalized.
 #' @seealso \code{\link{get_seats}}, \code{\link{sls}}
 #' @examples
 #' library(coalitions)
 #' library(dplyr)
 #' # get the latest survey for a sample of German federal election polls
-#' surveys <- get_latest(surveys_sample)
+#' surveys <- get_latest(surveys_sample) %>% ungroup() %>% slice(1)
 #' # redistribute the shares of 'others' parties and parties with a share of under 5\%
 #' surveys <- surveys %>% mutate(survey_redist = purrr::map(survey, redistribute))
 #' surveys$survey # results before redistribution